Fix for speed test decrease caused by loop.cycle
Noticed that the change in dbee0586a9350e9b3d3feb21758276c567359305 that added the `loop.cycle` option dramatically reduced render speed. For some reason, defining a function on the `_context` reduces the speed. I'm not sure why--would be good to also investigate why that happens, as this is more or less doing the same thing, just adding the function to a different object.

Conflicts:

	lib/tags.js